move GetInstallMechanism
